Build on Linux (GCC)

+++ src/paex_sine_c++.cpp
@@ -52,7 +52,7 @@ int main()
             std::vector<float> sampleData(400);
             for (size_t i = 0; i < sampleData.size(); ++i) {
                 sampleData[i] =
-                    static_cast<float>(sin(M_PI * 2 * i / sampleData.size()));
+                    static_cast<float>(sin(M_PI * 2.0 * static_cast<double>(i) / static_cast<double>(sampleData.size())));
             }
             return Sample(sampleData.begin(), sampleData.end());
         }
@@ -70,7 +70,7 @@ int main()
 
         void onTick(Mixer& audio) override
         {
-            audio.channel(0).set_frequency(sineWave.length() *
+            audio.channel(0).set_frequency(static_cast<float>(sineWave.length()) *
                                            notes[index++ % notes.size()]);
         }
